requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
f-string==0.1
idna==2.10
lxml==4.6.3
parse==1.19.0
python-dotenv==0.18.0
requests==2.25.1
schedule==1.1.0
soupsieve==2.2.1
urllib3==1.26.6
